move context types to dfu_context_handler
There is no reason to have them in the lower layer files.
This change makes it easier to re-use functions across the lower layer files.
Now we can delete the internal header file as well.
