+++ subsys/dfu/src/dfu_context_handler.c
@@ -18,6 +18,19 @@
 #define MODEM_DELTA_IMAGE 2
 
 LOG_MODULE_REGISTER(dfu_context_handler, CONFIG_DFU_CTX_LOG_LEVEL);
+
+static struct dfu_ctx dfu_ctx_mcuboot = {
+	.init  = dfu_ctx_mcuboot_init,
+	.write = dfu_ctx_mcuboot_write,
+	.done  = dfu_ctx_mcuboot_done,
+};
+
+static struct dfu_ctx dfu_ctx_modem = {
+	.init = dfu_ctx_modem_init,
+	.write = dfu_ctx_modem_write,
+	.done = dfu_ctx_modem_done,
+};
+
 static struct dfu_ctx *ctx;
 
 static inline void log_img_hdr(struct mcuboot_img_header_v2 *img_hdr)
